requirements.txt
instagrapi
openai
jsonschema
SQLAlchemy
python-dotenv
pandas
groq

